fixed tests

+++ src/main/java/meetup/event/service/TeamMemberServiceImpl.java
@@ -9,7 +9,6 @@ import meetup.event.mapper.TeamMemberMapper;
 import meetup.event.model.Event;
 import meetup.event.model.TeamMember;
 import meetup.event.model.TeamMemberRole;
-import meetup.event.repository.EventRepository;
 import meetup.event.repository.TeamMemberRepository;
 import meetup.exception.NotAuthorizedException;
 import meetup.exception.NotFoundException;
@@ -22,16 +21,12 @@ import java.util.List;
 @RequiredArgsConstructor
 public class TeamMemberServiceImpl implements TeamMemberService {
     private final TeamMemberRepository teamMemberRepository;
-    private final EventRepository eventRepository;
+    private final EventService eventService;
     private final TeamMemberMapper teamMemberMapper;
 
     @Override
     public TeamMemberDto addTeamMember(Long userId, NewTeamMemberDto newTeamMemberDto) {
-        Event event = eventRepository.findById(newTeamMemberDto.eventId()).orElseThrow(
-                () -> new NotFoundException(String.format("Event id = %d not found!", newTeamMemberDto.eventId())));
-        if (!event.getOwnerId().equals(userId)) {
-            checkTeamMemberManagerRoleInEvent(newTeamMemberDto.eventId(), userId);
-        }
+        checkTeamMemberManagerOrOwnerRoleInEvent(newTeamMemberDto.eventId(), userId);
         TeamMember teamMember = teamMemberRepository.save(teamMemberMapper.toTeamMember(newTeamMemberDto));
         log.info("Member id = '{}' was added to team event id = '{}' by user id = '{}'",
                 teamMember.getId().getUserId(), teamMember.getId().getEventId(), userId);
@@ -50,29 +45,20 @@ public class TeamMemberServiceImpl implements TeamMemberService {
 
     @Override
     public TeamMemberDto updateTeamMemberInEvent(Long userId, Long eventId, Long memberId, UpdateTeamMemberDto updateTeamMemberDto) {
-        existsEventById(eventId);
-        checkTeamMemberManagerRoleInEvent(eventId, userId);
+        checkTeamMemberManagerOrOwnerRoleInEvent(eventId, userId);
         TeamMember member = getTeamMember(eventId, memberId);
         teamMemberMapper.updateTeamMember(updateTeamMemberDto, member);
-        teamMemberRepository.save(member);
+        TeamMember upadatedTeamMember = teamMemberRepository.save(member);
         log.info("Member with id = '{}' was updated in team event id = '{}' by user id = '{}'", memberId, eventId, userId);
-        return teamMemberMapper.toTeamMemberDto(member);
+        return teamMemberMapper.toTeamMemberDto(upadatedTeamMember);
     }
 
     @Override
     public void deleteTeamMemberFromEvent(Long userId, Long eventId, Long memberId) {
-        existsEventById(eventId);
-        checkTeamMemberManagerRoleInEvent(eventId, userId);
+        checkTeamMemberManagerOrOwnerRoleInEvent(eventId, userId);
         TeamMember member = getTeamMember(eventId, memberId);
         teamMemberRepository.delete(member);
         log.info("Member with id = '{}' was deleted from team event id = '{}'", memberId, eventId);
-    }
-
-    private boolean existsEventById(Long eventId) {
-        if (!eventRepository.existsById(eventId)) {
-            throw new NotFoundException(String.format("Event id = %d not found!", eventId));
-        }
-        return true;
     }
 
     private TeamMember getTeamMember(Long eventId, Long memberId) {
@@ -81,10 +67,15 @@ public class TeamMemberServiceImpl implements TeamMemberService {
         );
     }
 
-    private void checkTeamMemberManagerRoleInEvent(Long eventId, Long memberId) {
-        TeamMember user = getTeamMember(eventId, memberId);
-        if (!user.getRole().equals(TeamMemberRole.MANAGER)) {
-            throw new NotAuthorizedException(String.format("User id = %d in event id = %d not Manager", memberId, eventId));
+    private void checkTeamMemberManagerOrOwnerRoleInEvent(Long eventId, Long memberId) {
+        Event event = eventService.getEventByEventId(eventId, memberId);
+        if (event == null) {
+            throw new NotFoundException("Event id = " + eventId + " not found!");
+        } else if (!event.getOwnerId().equals(memberId)) {
+            TeamMember user = getTeamMember(eventId, memberId);
+            if (!user.getRole().equals(TeamMemberRole.MANAGER)) {
+                throw new NotAuthorizedException(String.format("User id = %d in event id = %d not Manager", memberId, eventId));
+            }
         }
     }
 }
